+++ sketch-compose-core/src/commonMain/kotlin/com/github/panpf/sketch/target/ComposeTarget.common.kt
@@ -18,12 +18,10 @@
 package com.github.panpf.sketch.target
 
 import androidx.compose.ui.graphics.painter.Painter
-import com.github.panpf.sketch.ComponentRegistry
 import com.github.panpf.sketch.Image
 import com.github.panpf.sketch.Sketch
 import com.github.panpf.sketch.asSketchImage
 import com.github.panpf.sketch.request.ImageRequest
-import com.github.panpf.sketch.request.RequestInterceptor
 import com.github.panpf.sketch.request.internal.ComposeRequestDelegate
 import com.github.panpf.sketch.request.internal.RequestDelegate
 import com.github.panpf.sketch.resize.ComposeResizeOnDrawHelper
@@ -71,14 +69,4 @@ interface ComposeTarget : Target {
         }
         return null
     }
-
-    override fun getComponents(): ComponentRegistry? = ComponentRegistry.Builder().apply {
-        val toComposeBitmapInterceptor = getToComposeBitmapRequestInterceptor()
-        if (toComposeBitmapInterceptor != null) {
-            addRequestInterceptor(toComposeBitmapInterceptor)
-        }
-    }.build()
 }
-
-// TODO The time to convert ComposeBitmap is placed in ComposeTarget, so that the memory cache can store SkiaBitmapImage
-expect fun getToComposeBitmapRequestInterceptor(): RequestInterceptor?