+++ sketch-compose-core/src/nonAndroidTest/kotlin/com/github/panpf/sketch/compose/core/nonandroid/test/request/internal/SkiaBitmapToComposeBitmapRequestInterceptorTest.kt
@@ -1,5 +1,0 @@
-package com.github.panpf.sketch.compose.core.nonandroid.test.request.internal
-
-class SkiaBitmapToComposeBitmapRequestInterceptorTest {
-    // TODO test
-}