+++ CHANGELOG.md
@@ -12,6 +12,8 @@ Translations: [简体中文](CHANGELOG_zh.md)
 
 * fix: Fix the bug that the image cannot be loaded when the ImageView is attached to the window but
   the size is null due to padding. [#208](https://github.com/panpf/sketch/issues/208)
+* change: SkiaBitmapImage is now cached in the memory cache on non-Android platforms, not
+  ComposeBitmapImage.
 
 # 4.0.0-alpha08